feat: folders refactor
