+++ app/src/main/java/com/devfiveurjc/agendaly/fragments/TaskAddFragment.java
@@ -1,4 +1,4 @@
-package com.devfiveurjc.agendaly.ui;
+package com.devfiveurjc.agendaly.fragments;
 
 import android.os.Bundle;
 
@@ -16,7 +16,7 @@ import android.widget.Toast;
 import com.devfiveurjc.agendaly.R;
 import com.devfiveurjc.agendaly.crud.CRUDTask;
 import com.devfiveurjc.agendaly.databinding.FragmentTaskAddBinding;
-import com.devfiveurjc.agendaly.model.Task;
+import com.devfiveurjc.agendaly.models.Task;
 import com.devfiveurjc.agendaly.dialogs.DateTimeDialog;
 
 import java.util.Calendar;
