+++ app/src/main/java/com/devfiveurjc/agendaly/fragments/TaskEditFragment.java
@@ -1,4 +1,4 @@
-package com.devfiveurjc.agendaly.ui;
+package com.devfiveurjc.agendaly.fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -15,7 +15,7 @@ import androidx.navigation.fragment.NavHostFragment;
 import com.devfiveurjc.agendaly.R;
 import com.devfiveurjc.agendaly.crud.CRUDTask;
 import com.devfiveurjc.agendaly.databinding.FragmentTaskEditBinding;
-import com.devfiveurjc.agendaly.model.Task;
+import com.devfiveurjc.agendaly.models.Task;
 import com.devfiveurjc.agendaly.dialogs.DateTimeDialog;
 import com.google.android.material.textfield.TextInputEditText;
