+++ app/src/main/java/com/devfiveurjc/agendaly/fragments/TaskInfoFragment.java
@@ -1,4 +1,4 @@
-package com.devfiveurjc.agendaly.ui;
+package com.devfiveurjc.agendaly.fragments;
 
 import android.os.Bundle;
 
@@ -14,7 +14,7 @@ import android.widget.TextView;
 import com.devfiveurjc.agendaly.R;
 import com.devfiveurjc.agendaly.crud.CRUDTask;
 import com.devfiveurjc.agendaly.databinding.FragmentTaskInfoBinding;
-import com.devfiveurjc.agendaly.model.Task;
+import com.devfiveurjc.agendaly.models.Task;
 
 
 public class TaskInfoFragment extends Fragment {
