+++ app/src/main/res/layout/fragment_task_edit.xml
@@ -5,7 +5,7 @@
     android:id="@+id/taskAddConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.TaskEditFragment">
+    tools:context=".fragments.TaskEditFragment">
 
     <com.google.android.material.textfield.TextInputEditText
         android:id="@+id/taskEditEditTitle"
