+++ app/src/main/res/layout/fragment_task_info.xml
@@ -5,7 +5,7 @@
     android:id="@+id/taskInfoConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.TaskInfoFragment">
+    tools:context=".fragments.TaskInfoFragment">
 
     <TextView
         android:id="@+id/taskInfoTitle"
