+++ app/src/main/res/layout/fragment_task_list.xml
@@ -5,7 +5,7 @@
     android:id="@+id/taskListConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.TaskListFragment">
+    tools:context=".fragments.TaskListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/listRecyclerView"
