+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/TaskListFragment"
-        android:name="com.devfiveurjc.agendaly.ui.TaskListFragment"
+        android:name="com.devfiveurjc.agendaly.fragments.TaskListFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_task_list">
         <action
@@ -20,7 +20,7 @@
 
     <fragment
         android:id="@+id/TaskAddFragment"
-        android:name="com.devfiveurjc.agendaly.ui.TaskAddFragment"
+        android:name="com.devfiveurjc.agendaly.fragments.TaskAddFragment"
         android:label="@string/task_add_label"
         tools:layout="@layout/fragment_task_add">
         <action
@@ -30,7 +30,7 @@
 
     <fragment
         android:id="@+id/TaskInfoFragment"
-        android:name="com.devfiveurjc.agendaly.ui.TaskInfoFragment"
+        android:name="com.devfiveurjc.agendaly.fragments.TaskInfoFragment"
         android:label="@string/task_info_label"
         tools:layout="@layout/fragment_task_info">
         <action
@@ -43,7 +43,7 @@
 
     <fragment
         android:id="@+id/taskEditFragment"
-        android:name="com.devfiveurjc.agendaly.ui.TaskEditFragment"
+        android:name="com.devfiveurjc.agendaly.fragments.TaskEditFragment"
         android:label="@string/task_edit_label"
         tools:layout="@layout/fragment_task_edit" >
         <action
